Consolidate bike search components
