+++ app/javascript/packs/bikes_search/components/BikeSearchResult.js
