+++ app/javascript/packs/bikes_search/components/ExternalRegistrySearch.js
