+++ app/javascript/packs/bikes_search/components/ExternalRegistrySearchResult.js
