--- app/javascript/packs/partial_serial_search/index.js
@@ -1,17 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import ErrorBoundary from "@honeybadger-io/react";
-import PartialSerialSearch from "./components/PartialSerialSearch";
-import honeybadger from "../utils/honeybadger";
-
-document.addEventListener("DOMContentLoaded", () => {
-  const el = document.getElementById("js-partial-serial-search");
-  if (!el) { return; }
-
-  ReactDOM.render(
-    <ErrorBoundary honeybadger={honeybadger}>
-    <PartialSerialSearch interpretedParams={window.interpreted_params} />
-    </ErrorBoundary>,
-    el
-  );
-});
