+++ app/views/bikes/index.html.haml
@@ -1,6 +1,4 @@
-= javascript_pack_tag "partial_serial_search/index"
-= javascript_pack_tag "close_serial_search/index"
-= javascript_pack_tag "external_registry_search/index"
+= javascript_pack_tag "bikes_search/index"
 
 - stolenness_desc = { 'proximity' => 'nearby stolen', 'non' => 'non-stolen', 'stolen' => 'stolen', 'all' => 'all' }[@interpreted_params[:stolenness]]
